Main-queue-checker fix for deadlocking edge case
Previously, fcm_onMainQueue assumed that it was safe to dispatch_sync onto the main queue if we were not currently on it.

This is incorrect because there's a possibility for non-main-queue blocks to execute on the main thread: https://github.com/apple/swift-corelibs-libdispatch/commit/e64e4b962e1f356d7561e7a6103b424f335d85f6

dispatch_sync() always runs its block on the calling thread! So if the main queue calls dispatch_sync() to a background queue, that background-queue work will execute on the main thread — but not the main queue.

If dispatch_sync() onto the main thread is called from such a context, it'll deadlock.

To ensure that it's safe to dispatch_sync() to the main thread, we need the dispatch_get_specific() check AND (! NSThread.isMainThread).

I had this deadlock happen reliably during snapshot changes on UITableViewDiffableDataSource, and this change fixed it.